*Created Reset Password pages and Functionality -Created validatePasswordEmail api call function to validate the email provided by the user for password resetting -Created resetPassword api call function to manage the creation of the new password -Made SetPasswordForm a reusable compnent by passing props to it to manage both the restoration of a password and the creation of one
*Other Enhancments
-Created 404 custom page
-Created a fix type for the user admin table because the other type that was build upon
userSchema does not have in account that displayed user COULD not have his password setted.
Now usersResponseSchema has adminTableUserSchema as base and userResponseSchema accepts the unique token in the api response
-RegisterView now formats the inputted rut in real time

*Fixed errors reagarding AdminConfirmedTable
-When a user was not confirmed, the api call was made using the user id as a parameter instead of the
admin confirmation token. Fixed this error both in the backend and in the frontend to have the token in the
getUserById endpoint response.